requirements.txt
selenium==4.21.0
pytz
python-dotenv
packaging
